+++ scripts/wrappers/effectIconWrappers.js
@@ -6,6 +6,7 @@ const feature_id = 'EffectIcons';
 
 export function registerEffectIconWrappers() {
   unregisterWrapper(feature_id, 'Token.prototype.drawEffects');
+  unregisterWrapper(feature_id, 'CombatTracker.prototype.getData');
   if (!FEATURE_CONTROL[feature_id]) return;
 
   if (
@@ -24,6 +25,15 @@ export function registerEffectIconWrappers() {
   } else if (TVA_CONFIG.displayEffectIconsOnHover) {
     registerWrapper(feature_id, 'Token.prototype.drawEffects', _drawEffects_hoverOnly, 'WRAPPER');
   }
+
+  if (TVA_CONFIG.disableEffectIcons || TVA_CONFIG.filterCustomEffectIcons) {
+    registerWrapper(
+      feature_id,
+      'CombatTracker.prototype.getData',
+      _combatTrackerGetData,
+      'WRAPPER'
+    );
+  }
 }
 
 async function _drawEffects_hoverOnly(wrapped, ...args) {
@@ -36,6 +46,42 @@ async function _drawEffects_fullReplace(...args) {
   this.effects.removeChildren().forEach((c) => c.destroy());
   this.effects.bg = this.effects.addChild(new PIXI.Graphics());
   this.effects.overlay = null;
+}
+
+async function _combatTrackerGetData(wrapped, ...args) {
+  let data = await wrapped(...args);
+
+  if (data && data.combat && data.turns) {
+    const combat = data.combat;
+    for (const turn of data.turns) {
+      const combatant = combat.combatants.find((c) => c.id === turn.id);
+      if (combatant) {
+        if (TVA_CONFIG.disableEffectIcons) {
+          turn.effects = new Set();
+        } else if (TVA_CONFIG.filterEffectIcons) {
+          const restrictedEffects = _getRestrictedEffects(combatant.token);
+
+          // Copied from CombatTracker.getData(...)
+          turn.effects = new Set();
+          if (combatant.token) {
+            combatant.token.effects.forEach((e) => turn.effects.add(e));
+            if (combatant.token.overlayEffect) turn.effects.add(combatant.token.overlayEffect);
+          }
+
+          // modified to filter restricted effects
+          if (combatant.actor) {
+            for (const effect of combatant.actor.temporaryEffects) {
+              if (effect.statuses.has(CONFIG.specialStatusEffects.DEFEATED)) {
+              } else if (effect.icon && !restrictedEffects.includes(effect.name ?? effect.label))
+                turn.effects.add(effect.icon);
+            }
+          }
+          // end of copy
+        }
+      }
+    }
+  }
+  return data;
 }
 
 async function _drawEffects(...args) {
@@ -55,13 +101,7 @@ async function _drawEffects(...args) {
   // Modified from the original token.drawEffects
   if (TVA_CONFIG.displayEffectIconsOnHover) this.effects.visible = this.hover;
   if (tokenEffects.length || actorEffects.length) {
-    let restrictedEffects = TVA_CONFIG.filterIconList;
-    if (TVA_CONFIG.filterCustomEffectIcons) {
-      const mappings = getAllEffectMappings({
-        actor: this.actor ? this.actor : this.document,
-      });
-      if (mappings) restrictedEffects = restrictedEffects.concat(mappings.map((m) => m.expression));
-    }
+    const restrictedEffects = _getRestrictedEffects(this.document);
     actorEffects = actorEffects.filter((ef) => !restrictedEffects.includes(ef.name ?? ef.label));
     tokenEffects = tokenEffects.filter(
       // check if it's a string here
@@ -95,4 +135,13 @@ async function _drawEffects(...args) {
   this.effects.overlay = await this._drawOverlay(overlay.src, overlay.tint);
   this._refreshEffects();
   this.effects.renderable = true;
+}
+
+function _getRestrictedEffects(tokenDoc) {
+  let restrictedEffects = TVA_CONFIG.filterIconList;
+  if (TVA_CONFIG.filterCustomEffectIcons) {
+    const mappings = getAllEffectMappings(tokenDoc);
+    if (mappings) restrictedEffects = restrictedEffects.concat(mappings.map((m) => m.expression));
+  }
+  return restrictedEffects;
 }